+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [1.5.4] - 2018-05-07
+### Added
+- added tachyons css library
+- added new listing page for fac/staff by category showing cards
+
 ## [1.5.3] - 2017-12-07
 ### Changed
 - now sorting categories by slug when using 'By Category' display style